services/bot/requirements.txt
aiofiles==22.1.0
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==22.1.0
chardet==4.0.0
charset-normalizer==2.1.1
discord-components @ git+https://github.com/kiki7000/discord.py-components.git@master
discord.py==1.7.3
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
nats-py==2.1.7
python-dotenv==0.21.0
typing_extensions==4.3.0
yarl==1.8.1
